Add files via upload
Csv with Parse: collects data from the Keithley, writes the data into a csv file (one column for voltage, one for time and one for voltage but without units). this should be continuously updating.
Live plotting: should take the data from the continuously updating csv and make a graph using the voltage without units for the y and time for the x axis.
LivePlottingfromCSV: this was an attempt to have both the Csv with parse and the live plotting in the same file, but it doesn't work.
Read csv: is Csv with parse but it only have the time and voltage with units.